+++ SAMPLE/ComponentRouting.Maui.Sample/Presenters/Tabs/HomePage.xaml.cs
@@ -9,6 +9,7 @@ public partial class HomePage : ContentPage, Presenter, OverlayHost
     private Func<Task>? startWizard;
     private Func<Task>? showLoading;
     private Func<Task>? hideLoading;
+    private Func<Task>? closeAllPopups;
     private Func<Task>? showSnackbar;
     private Func<Task>? countSnackbars;
 
@@ -25,6 +26,7 @@ public partial class HomePage : ContentPage, Presenter, OverlayHost
         Func<Task> startWizard,
         Func<Task> showLoading,
         Func<Task> hideLoading,
+        Func<Task> closeAllPopups,
         Func<Task> showSnackbar,
         Func<Task> countSnackbars)
     {
@@ -33,6 +35,7 @@ public partial class HomePage : ContentPage, Presenter, OverlayHost
         this.startWizard = startWizard;
         this.showLoading = showLoading;
         this.hideLoading = hideLoading;
+        this.closeAllPopups = closeAllPopups;
         this.showSnackbar = showSnackbar;
         this.countSnackbars = countSnackbars;
     }
@@ -59,6 +62,7 @@ public partial class HomePage : ContentPage, Presenter, OverlayHost
     private async void HandleStartWizardClicked(object? sender, EventArgs e) => await Invoke(startWizard);
     private async void HandleShowLoadingClicked(object? sender, EventArgs e) => await Invoke(showLoading);
     private async void HandleHideLoadingClicked(object? sender, EventArgs e) => await Invoke(hideLoading);
+    private async void HandleCloseAllPopupsClicked(object? sender, EventArgs e) => await Invoke(closeAllPopups);
     private async void HandleShowSnackbarClicked(object? sender, EventArgs e) => await Invoke(showSnackbar);
     private async void HandleCountSnackbarsClicked(object? sender, EventArgs e) => await Invoke(countSnackbars);